Move gameplay clock to own class
